fix(publish): offer both RC top-up routes, and fix href button alignment
Two problems with the out-of-RC alert.

The label overflowed its button. Button only applies flex centering when
an icon is present; a <button> centres its content natively but an <a>
does not, so any href Button without an icon rendered as an inline
anchor with the label sitting on the text baseline outside the padded
box. Fixed in the component rather than at the call site, because it hit
every href Button without an icon, PointsTopupCta included.

The alert also offered only a hardcoded Boost link, bypassing the
top-up product entirely. Ecency already supports spending Points on a
short-term RC-only delegation to your own account, which is the direct
fix for being out of RC, and buying Points by card when the balance is
short. The alert now offers both, and falls back to the Boost page when
the RC top-up flag is off.

That fallback logic already existed in RcPrecheckBanner, so rather than
copy it the decision moved into a shared useRcTopupAction hook that both
the pre-publish warning and the post-failure alert use. The banner's
behaviour is unchanged; it just no longer owns the only copy.

buildRcTopUpUrl is gone with its specs: it hardcoded the boost product
and nothing calls it now.
