+++ apps/web/src/features/shared/rc-topup/index.ts
@@ -1,1 +1,2 @@
 export * from "./rc-topup-dialog";
+export * from "./use-rc-topup-action";
